+++ lib/src/templates.ts
@@ -1,5 +1,4 @@
 import { JSONSchema7TypeName } from 'json-schema';
-import Section from './components/Fields/Section';
 import VernaJSONSchemaType from './types/rjsf';
 
 const stringDefinition = (type: JSONSchema7TypeName = 'string'): VernaJSONSchemaType => {
@@ -43,15 +42,10 @@ const sectionParametersSchema = (): VernaJSONSchemaType => ({
 
 const defaultVernaWidgets = {};
 
-const defaultObjectFieldTemplate = {
-  section: Section,
-};
-
 export {
   stringDefinition,
   sectionDefinition,
   defaultVernaWidgets,
-  defaultObjectFieldTemplate,
   defaultWidgetProps,
   sectionParametersSchema,
 };